Store outgoing trust in ProjectTrusts array
The Dynamics API currently does not correctly store/return the outgoing
trust in the ProjectAcademies array. This means it cannot be accessed in
any future pages.

To allow for accessing the outgoing trust of the project - it is now
stored in index 1 of the ProjectTrusts array in the Project model