Add custom Google Tag Manager event when project template downloaded
We want to know which sections have been completed when the project template is downloaded. This isn't captured by GTM out-of-the-box, so we need to add a custom event.

Add data attributes to the section statuses and generate button in order to build and dispatch the custom event.

This also reconfigures Webpack so that we can add multiple JavaScript files.