Convert SubmitOutGoingTrustAcademies to Razor Pages POST method
We use Razor Pages for the majority of the app, and want to convert the remaining actions in the TransfersController. 

Refactor the SubmitOutGoingTrustAcademies GET action from Razor Views to a Razor Pages POST method. 

This also updates the method to use ModelState instead of passing error messages around with TempData, uses model binding for the submitted academy IDs, and uses the validation tag helper for error messages in order to DRY up the view a bit.