Convert TrustName action to Razor Pages
We use Razor Pages for the majority of the app, and want to convert the remaining actions in the TransfersController. 

Refactor the TrustName action from Razor Views to Razor Pages. 

This also uses model binding for the search query, and uses the validation tag helper for error messages in order to DRY up the views a bit.

We still need TempData for the Error messages because validation is performed on the subsequent TrustSearch page, but this is now mapped to ModelState.