Update the HTML for summary lists
We currently add an empty `dd` tag in the place of an action in summary lists when there are a mix of rows with and without actions. We do this both manually, and through the `gds-key-value` tag helper when `show-action` is true.

The recommended way of handling this is now to apply the `govuk-summary-list__row--no-action` class which inserts an empty cell pseudo-element instead.