Clear session cookie when user signs out of Azure AD
We can use SignOutAsync to sign out of Azure AD if we specify the scheme as OpenIDConnect (default value).

We also need to clear the .NET Core authentication session cookie, which can be achieved by calling SignOutAsync with the cookie name.

When a user signs out they will see the Azure AD sign-out page followed by a "You have signed out" page on the service.