Remove the reason for transfer question
Currently, there is some repetition in the features of the transfer section, which we want to reduce down to a single "what is the reason for this transfer?" question.

We are going to use the "Who initiated this transfer?" data for this, so removing the current "reason" question. This will be followed up by an update to the "Who initiated this transfer?" page.