Fix date validation
We want the expected date of transfer to always be the first of the month.

I was previously setting the day property unconditionally, which meant that the validation was returning an "invalid date" message, instead of a "required field" message when the other fields were empty.

To fix - conditionally set day only if month and year are also set.