Update the template file name
We want the project file to be more intuitive, which we can achieve by including the outgoing trust name (in addition to the incoming trust name) and moving the project reference to the beginning.

Add string helpers to hyphenate sentences and remove non-alphanumeric characters so that trust names can be formatted as they are in the updated design.

Also, add missing string helper tests and DRY up some duplication between the download handler methods.