Only show the field missing error when another field has a value
The expected date of transfer doesn't have a day field, so we only want to show the missing field error message when either the month or year field has a value.

Currently, we are setting the day value when both month and year have a value, but this showing a field missing error. Update the conditional to fix.